desktop-client/requirements.txt
SpeechRecognition
pyttsx3
pillow
pyaudio